pyproject.toml
[tool.poetry]
name = "romaprivatebot"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
pyQiwiP2P = "^2.0.1"
python-dotenv = "0.17.1"
pydantic = "1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
